requirements.txt
asgiref==3.8.1
Django==5.2
django-filter==25.1
django-modeltranslation==0.19.14
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
Markdown==3.8
pillow==11.2.1
psycopg2-binary==2.9.10
PyJWT==2.9.0
sqlparse==0.5.3
tzdata==2025.2
